move status types
