+++ internal/agent/reporter.go
@@ -23,6 +23,7 @@ import (
 	"github.com/dagu-org/dagu/internal/dag"
 	"github.com/dagu-org/dagu/internal/dag/scheduler"
 	"github.com/dagu-org/dagu/internal/logger"
+	"github.com/dagu-org/dagu/internal/persistence/history"
 	"github.com/dagu-org/dagu/internal/persistence/model"
 	"github.com/jedib0t/go-pretty/v6/table"
 )
@@ -50,7 +51,7 @@ func newReporter(sender Sender, lg logger.Logger) *reporter {
 
 // reportStep is a function that reports the status of a step.
 func (r *reporter) reportStep(
-	dAG *dag.DAG, status *model.Status, node *scheduler.Node,
+	dAG *dag.DAG, status *history.Status, node *scheduler.Node,
 ) error {
 	nodeStatus := node.State().Status
 	if nodeStatus != scheduler.NodeStatusNone {
@@ -74,7 +75,7 @@ func (r *reporter) reportStep(
 }
 
 // report is a function that reports the status of the scheduler.
-func (r *reporter) report(status *model.Status, err error) {
+func (r *reporter) report(status *history.Status, err error) {
 	var buf bytes.Buffer
 	_, _ = buf.Write([]byte("\n"))
 	_, _ = buf.Write([]byte("Summary ->\n"))
@@ -87,7 +88,7 @@ func (r *reporter) report(status *model.Status, err error) {
 
 // send is a function that sends a report mail.
 func (r *reporter) send(
-	dAG *dag.DAG, status *model.Status, err error,
+	dAG *dag.DAG, status *history.Status, err error,
 ) error {
 	if err != nil || status.Status == scheduler.StatusError {
 		if dAG.MailOn != nil && dAG.MailOn.Failure {
@@ -117,7 +118,7 @@ func (r *reporter) send(
 	return nil
 }
 
-func renderSummary(status *model.Status, err error) string {
+func renderSummary(status *history.Status, err error) string {
 	t := table.NewWriter()
 	var errText string
 	if err != nil {
